requirements.txt
fastapi
uvicorn
motor
python-dotenv
aiosmtplib
email-validator
python-multipart
python-jose[cryptography]
